+++ domain/src/main/java/woowacourse/shopping/domain/repository/ShoppingCartRepository.kt
@@ -5,16 +5,16 @@ import woowacourse.shopping.domain.util.WoowaResult
 
 interface ShoppingCartRepository {
     fun fetchAll(callback: (WoowaResult<List<CartProduct>>) -> Unit)
-    fun delete(callback: (WoowaResult<Boolean>) -> Unit, id: Int)
+    fun delete(callback: (WoowaResult<Boolean>) -> Unit, id: Long)
     fun insert(
         callback: (WoowaResult<Long>) -> Unit,
-        productId: Int,
+        productId: Long,
         quantity: Int = 1,
     )
 
     fun update(
         callback: (WoowaResult<Boolean>) -> Unit,
-        productId: Int,
+        productId: Long,
         updatedQuantity: Int,
     )
 }
